// src/shared/copilot/types.ts
export interface DomainSuggestion {
  title: string;
  message: string;
}